# Reduce memory for suite configs with lots of defaults