--- a/lib/cylc/batch_sys_handlers/lsf.py
+++ b/lib/cylc/batch_sys_handlers/lsf.py
@@ -44,7 +44,8 @@ class LSFHandler(object):
         directives['-J'] = job_conf['suite name'] + '.' + job_conf['task id']
         directives['-o'] = job_file_path + ".out"
         directives['-e'] = job_file_path + ".err"
-        directives.update(job_conf['directives'])
+        for key, value in job_conf['directives'].items():
+            directives[key] = value
         lines = []
         for key, value in directives.items():
             if value: