--- a/lib/parsec/OrderedDict.py
+++ b/lib/parsec/OrderedDict.py
@@ -18,6 +18,7 @@
 
 """Ordered Dictionary data structure used extensively in cylc."""
 
+
 try:
     # first try the fast ordereddict C implementation.
     # DOWNLOAD: http://anthon.home.xs4all.nl/Python/ordereddict/
@@ -43,32 +44,31 @@ class OrderedDictWithDefaults(OrderedDict):
 
     """
 
-    def __contains_default__(self, key):
-        """Make sure "key in foo" works with our defaults."""
-        if key in getattr(self, "defaults", {}):
-            return True
-        return OrderedDict.__contains__(self, key)
+    def __init__(self, *args, **kwargs):
+        """Allow a defaults argument."""
+        self._allow_contains_default = True
+        super(OrderedDictWithDefaults, self).__init__(*args, **kwargs)
 
     def __getitem__(self, key):
         """Override to look in our special .defaults attribute, if it exists."""
         try:
             return OrderedDict.__getitem__(self, key)
         except KeyError:
-            if hasattr(self, 'defaults'):
-                return self.defaults[key]
+            if hasattr(self, 'defaults_'):
+                return self.defaults_[key]
             raise
 
     def __setitem__(self, *args, **kwargs):
         """Make sure that we don't set the default value!"""
-        self.__contains__ = OrderedDict.__contains__
+        self._allow_contains_default = False
         return_value = OrderedDict.__setitem__(self, *args, **kwargs)
-        self.__contains__ = self.__contains_default__
+        self._allow_contains_default = True
         return return_value
 
     def keys(self):
         """Include the default keys, after the list of actually-set ones."""
         keys = list(self)
-        for key in getattr(self, 'defaults', []):
+        for key in getattr(self, 'defaults_', []):
             if key not in keys:
                 keys.append(key)
         return keys
@@ -96,6 +96,12 @@ class OrderedDictWithDefaults(OrderedDict):
         for k in self.keys():
             yield (k, self[k])
 
+    def __contains__(self, key):
+        if self._allow_contains_default:
+            if key in getattr(self, "defaults_", {}):
+                return True
+        return OrderedDict.__contains__(self, key)
+
     def __nonzero__(self):
         """Include any default keys in the nonzero calculation."""
         return bool(self.keys())
@@ -107,10 +113,8 @@ class OrderedDictWithDefaults(OrderedDict):
         for key in non_default_keys:
             non_default_items.append((key, self[key]))
         default_items = []
-        for key in getattr(self, 'defaults', []):
+        for key in getattr(self, 'defaults_', []):
             if key not in non_default_keys:
                 default_items.append((key, self[key]))
-        repr_map = {"": non_default_items, "defaults": default_items}
+        repr_map = {"": non_default_items, "defaults_": default_items}
         return "<" + type(self).__name__ + "(" + repr(repr_map) + ")>\n"
-
-    __contains__ = __contains_default__