--- a/lib/parsec/OrderedDict.py
+++ b/lib/parsec/OrderedDict.py
@@ -32,3 +32,57 @@ except ImportError:
         # then try the pre-2.7 backport from ActiveState
         # (packaged with cylc)
         from OrderedDictCompat import OrderedDict
+
+
+class OrderedDictWithDefaults(OrderedDict):
+
+    """Subclass to provide defaults fetching capability.
+
+    Note that defining a '__missing__' method would work for foo[key],
+    but doesn't for foo.get(key).
+
+    """
+
+    def __contains__(self, key):
+        """Make sure "key in foo" works with our defaults."""
+        return key in self.keys()
+
+    def __getitem__(self, key):
+        """Override to look in our special .defaults attribute, if it exists."""
+        try:
+            return OrderedDict.__getitem__(self, key)
+        except KeyError:
+            if hasattr(self, 'defaults'):
+                return self.defaults[key]
+            raise
+
+    def keys(self):
+        """Include the default keys, after the list of actually-set ones."""
+        keys = list(self)
+        for key in getattr(self, 'defaults', []):
+            if key not in keys:
+                keys.append(key)
+        return keys
+
+    def values(self):
+        """Return a list of values, including default ones."""
+        return [self[key] for key in self.keys()]
+
+    def items(self):
+        """Return key-value pairs, including default ones."""
+        return [(key, self[key]) for key in self.keys()]
+
+    def iterkeys(self):
+        """Include default keys - no memory saving over .keys()."""
+        for k in self.keys():
+            yield k
+
+    def itervalues(self):
+        """Include default values - no memory saving over .values()."""
+        for k in self.keys():
+            yield self[k]
+
+    def iteritems(self):
+        """Include default key-value pairs - no memory saving over .items()"""
+        for k in self.keys():
+            yield (k, self[k])