--- a/lib/parsec/config.py
+++ b/lib/parsec/config.py
@@ -20,7 +20,7 @@ import os, sys, re
 from parsec.fileparse import parse, FileNotFoundError
 from parsec.util import printcfg
 from parsec.validate import validate, check_compulsory, expand, validator
-from parsec.OrderedDict import OrderedDict
+from parsec.OrderedDict import OrderedDict, OrderedDictWithDefaults
 from parsec.util import replicate, itemstr
 from parsec.upgrade import UpgradeError
 import cylc.flags
@@ -45,8 +45,8 @@ class config( object ):
     def __init__( self, spec, upgrader=None, write_proc=False,
             tvars=[], tvars_file=None ):
 
-        self.sparse = OrderedDict()
-        self.dense = OrderedDict()
+        self.sparse = OrderedDictWithDefaults()
+        self.dense = OrderedDictWithDefaults()
         self.upgrader = upgrader
         self.tvars = tvars
         self.tvars_file = tvars_file