… lib/parsec/util.py
… tests/inheritance/00-namespace-list.t