This reduces parsec memory usage for configurations with lots of default settings. For a suite like this:
```ini
#!jinja2
[scheduling]
    [[dependencies]]
        graph = shutdown => LOTS
[runtime]
    [[shutdown]]
        command scripting = "cylc shutdown $CYLC_SUITE_NAME; sleep 60"
    [[LOTS]]
        command scripting = echo MY_NUM
    {%- for i in range(10000) %}
    {%- set i_label = "%05d" % i %}
    [[foo_{{i_label}}]]
        inherit = LOTS
        [[[environment]]]
            MY_NUM = {{i}}
    {%- endfor %}
```
it reduces total `cylc run` memory use by a factor of 2. It isn't quite as dramatic for some suites, though. It doesn't appear to significantly impact CPU usage.

When you have a parsec section with lots of defaults, the basic premise is that instead of copying them into the main configuration, you add a redirect reference to the defaults dictionary. Then, when a key isn't found, you just go to the defaults dict and look it up.

I freely admit that the implementation is a bit yucky. Suggestions welcome.

This is related to #1222.

I still need to test this with the "fast ordereddict C implementation" recipe.

@hjoliver, please review. I've put it in 'soon' in case we want to leave this out for now.