@hjoliver 2015-06-16
> I still need to test this with the "fast ordereddict C implementation" recipe.

I have not bothered to use this myself for quite some time.  We should probably check to see if provides any significant advantage (I think it used to, but that was very pre-parsec) and install it properly if it does, or else remove it.