@hjoliver 2015-06-16
@benfitzpatrick - I'm happy with this, and the implementation doesn't look too "yucky" to me.   Two questions though: (1) if this eliminates all duplication of default values, is it surprising that it only halves memory use in very large suites?  (2) can this be extended to treat inherited items like defaults?