@benfitzpatrick 2015-06-16
1) The defaults were the largest area, but others are pretty significant. The defaults memory usage hasn't been completely removed either.

Here's some memory benchmarking for the above suite (note: starting memory use seems to vary a bit (import order?))

location | master (KiB) | #1500 branch (KiB)
----------- | ------------ | -------------
SCHEDULER CONFIGURE | 17904 | 19976
BEFORE CONFIGURE SUITE | 18204 | 22076
INIT CONFIG | 18204 | 22076
BEFORE GETSUITECFG | 18208 | 22076
AFTER GETSUITECFG | 57288 | 60896
BEFORE GETSPARSE | 57288 | 60896
AFTER GETSPARSE | 57288 | 60896
BEFORE FAMILY TREE | 75752 | 81108
AFTER FAMILY TREE | 76660 | 87440
BEFORE INHERITANCE | 76660 | 87440
AFTER INHERITANCE | 76660 | 90540
BEFORE DEFAULTS EXPANSION | 76660 | 90644
AFTER DEFAULTS EXPANSION | 355764 | 118040
BEFORE LOAD GRAPH | 355764 | 118040
AFTER LOAD GRAPH | 397432 | 160636
END INIT CONFIG | 397516 | 160724
AFTER CONFIGURE SUITE | 398228 | 161548
BEFORE LOAD TASKS | 398260 | 161580
AFTER LOAD TASKS | 629100 | 389612
END SCHEDULER CONFIGURE | 629100 | 389616
BEGIN RUN WHILE LOOP | 629100 | 389616

Those numbers were produced with the following diff:

```diff
diff --git a/lib/cylc/config.py b/lib/cylc/config.py
index 6c101fa..5cd41fb 100644
--- a/lib/cylc/config.py
+++ b/lib/cylc/config.py
@@ -122,6 +122,7 @@ class config( object ):
                  is_restart=False, is_reload=False, write_proc=True,
                  vis_start_string=None, vis_stop_string=None):
 
+        printmem("INIT CONFIG")
         self.suite = suite  # suite name
         self.fpath = fpath  # suite definition
         self.fdir  = os.path.dirname(fpath)
@@ -173,10 +174,14 @@ class config( object ):
         self.feet = []
 
         # parse, upgrade, validate the suite, but don't expand with default items
+        printmem("BEFORE GETSUITECFG")
         self.pcfg = get_suitecfg( fpath, force=is_reload,
                 tvars=template_vars, tvars_file=template_vars_file,
                 write_proc=write_proc )
+        printmem("AFTER GETSUITECFG")
+        printmem("BEFORE GETSPARSE")
         self.cfg = self.pcfg.get(sparse=True)
+        printmem("AFTER GETSPARSE")
 
         if self._cli_initial_point_string is not None:
             self.cfg['scheduling']['initial cycle point'] = (
@@ -261,9 +266,13 @@ class config( object ):
         # check var names before inheritance to avoid repetition
         self.check_env_names()
 
+        printmem("BEFORE FAMILY TREE")
         # do sparse inheritance
         self.compute_family_tree()
+        printmem("AFTER FAMILY TREE")
+        printmem("BEFORE INHERITANCE")
         self.compute_inheritance()
+        printmem("AFTER INHERITANCE")
 
         #self.print_inheritance() # (debugging)
 
@@ -271,7 +280,9 @@ class config( object ):
         self.filter_env()
 
         # now expand with defaults
+        printmem("BEFORE DEFAULTS EXPANSION")
         self.cfg = self.pcfg.get( sparse=False )
+        printmem("AFTER DEFAULTS EXPANSION")
 
         # after the call to init_cyclers, we can start getting proper points.
         init_cyclers(self.cfg)
@@ -465,7 +476,9 @@ class config( object ):
 
         self.process_directories()
 
+        printmem("BEFORE LOAD GRAPH")
         self.load_graph()
+        printmem("AFTER LOAD GRAPH")
 
         self.compute_runahead_limits()
 
@@ -629,6 +642,7 @@ class config( object ):
                     print >> sys.stderr, '  %s => %s' % e
                 raise SuiteConfigError('ERROR: cyclic dependence detected '
                                        '(graph the suite to see back-edges).')
+        printmem("END INIT CONFIG")
  
     def dequote(self, s):
         """Strip quotes off a string."""
@@ -2106,3 +2120,13 @@ class config( object ):
     def describe(self, name):
         """Return title and description of the named task."""
         return self.taskdefs[name].describe()
+
+import subprocess
+
+def getmem():
+    proc = subprocess.Popen(["ps", "h", "-orss", str(os.getpid())], stdout=subprocess.PIPE)
+    return int(proc.communicate()[0])
+
+
+def printmem(message):
+    print message, ":", getmem()
diff --git a/lib/cylc/scheduler.py b/lib/cylc/scheduler.py
index 10c8385..1e55ba3 100644
--- a/lib/cylc/scheduler.py
+++ b/lib/cylc/scheduler.py
@@ -41,7 +41,7 @@ from cylc.job_file import JOB_FILE
 from cylc.suite_host import get_suite_host
 from cylc.owner import user
 from cylc.version import CYLC_VERSION
-from cylc.config import config
+from cylc.config import config, printmem
 from parsec.util import printcfg
 from shutil import copy as shcopy, copytree, rmtree
 from copy import deepcopy
@@ -181,6 +181,7 @@ class scheduler(object):
         self.parse_commandline()
 
     def configure(self):
+        printmem("SCHEDULER CONFIGURE")
         SuiteProcPool.get_inst()
 
         # Get control and info commands.
@@ -210,7 +211,10 @@ class scheduler(object):
             'reload_suite',
             'add_prerequisite'
         ]
+
+        printmem("BEFORE CONFIGURE SUITE")
         self.configure_suite()
+        printmem("AFTER CONFIGURE SUITE")
 
         # Remotely accessible suite identifier
         self.suite_id = identifier(self.suite, self.owner)
@@ -245,7 +249,9 @@ class scheduler(object):
         self.request_handler.start()
 
         self.old_user_at_host_set = set()
+        printmem("BEFORE LOAD TASKS")
         self.load_tasks()
+        printmem("AFTER LOAD TASKS")
 
         self.state_dumper.set_cts(self.initial_point, self.final_point)
         self.configure_suite_environment()
@@ -286,6 +292,7 @@ class scheduler(object):
         self.nudge_timer_start = None
         self.nudge_timer_on = False
         self.auto_nudge_interval = 5  # seconds
+        printmem("END SCHEDULER CONFIGURE")
 
     def process_command_queue(self):
         queue = self.command_queue.get_queue()
@@ -889,6 +896,7 @@ class scheduler(object):
             'cylc']['event hooks']['abort if startup handler fails']
         self.run_event_handlers('startup', abort, 'suite starting')
 
+        printmem("BEGIN RUN WHILE LOOP")
         proc_pool = SuiteProcPool.get_inst()
         while True:  # MAIN LOOP
             # PROCESS ALL TASKS whenever something has changed that might
@@ -1004,6 +1012,7 @@ class scheduler(object):
                 )
             time.sleep(1)
 
+        printmem("END MAIN LOOP")
         # END MAIN LOOP
 
     def update_state_summary(self):
```