@hjoliver 2015-06-16
Nice. It might be worth adding this permanently, to run in debug mode, or a new profiling mode.